GH-2: Documented using the script as an executable in your path.  Renamed `setLogging` option.  Fixed unused global variable.
