+++ README.md
@@ -10,11 +10,13 @@ Before you can use this package, you will need to install its dependencies using
 
 ## Usage
 
-To launch the router, use a command like:
+The most basic way to launch the route is using a command like:
 
-`node .`
+`node launcher.js`
 
-If you launch the command without any arguments (or with the `--help` option), usage information will be displayed.
+You can also install flocking-midi-router as a global command available in your path by running `npm link` (or `yarn
+link`).  This will add a command named `flocking-midi-router` to your path.  If you launch either variation without any
+arguments (or with the `--help` option), usage information will be displayed.
 
 
 ### Listing Ports
@@ -22,7 +24,7 @@ If you launch the command without any arguments (or with the `--help` option), u
 To use this command, you will need to know part or all of the name of the attached input and output devices.  You can
 display all detected inputs and outputs using a commmand like:
 
-`node . -l`
+`flocking-midi-router -l`
 
 A list of outputs like the following will be displayed:
 
@@ -62,12 +64,12 @@ To route ports, pass in one or more inputs, and one or more outputs, as in the f
 the above example:
 
 ```
-node . -i Oxygen -o "MIDI Monitor"
+flocking-midi-router -i Oxygen -o "MIDI Monitor"
 ```
 
 Note the quotation marks around the output.  You can also pass in space-delimited arrays, as in the following example
 (which assumes you have inputs named `input1` and `input 2`, and outputs named `output 1` and `output2`:
 
 ```
-node . -i input1 "input 2" -o "output 1" output2
+flocking-midi-router -i input1 "input 2" -o "output 1" output2
 ```
